Moved ics file parsing method to TimetableServiceImpl class
