+++ src/main/java/com/team701/buddymatcher/services/timetable/TimetableService.java
@@ -2,6 +2,10 @@ package com.team701.buddymatcher.services.timetable;
 
 import com.team701.buddymatcher.domain.timetable.Timetable;
 
+import java.io.File;
+import java.util.List;
+
 public interface TimetableService {
     Timetable retrieve(String userId);
+    List<String> getCalInfoFromIcs(File file) throws Exception;
 }
